+++ packages/extension-ltree/src/exports/index-types.ts
@@ -1,0 +1,2 @@
+export type { GistIndexOptions, LtreeIndexTypes } from "../types/index-types";
+export { ltreeIndexTypes } from "../types/index-types";